Fix warnings
git-svn-id: https://svn.code.sf.net/p/tass64/code/trunk@2427 111d7451-93fa-4e81-90d4-8039f76616e9

+++ anonsymbolobj.h
@@ -26,12 +26,12 @@ struct file_list_s;
 
 typedef struct Anonsymbol {
     Obj v;
-    int32_t count;
+    ssize_t count;
 } Anonsymbol;
 
 extern void anonsymbolobj_init(void);
 
-extern Anonsymbol *new_anonsymbol(int32_t);
+extern Anonsymbol *new_anonsymbol(ssize_t);
 
 static inline Anonsymbol *ref_anonsymbol(Anonsymbol *v1) {
     v1->v.refcount++; return v1;
